Use hooks instead of context render props in withAppProvider
This didn't play nice with how overrode isServer in
server-side-rendering tests due to funky mocking. Splitting the mocks
out into separate files fixes this. Strictly speaking this only caused
problems in Avatar as it is a functional component, but I figured lets
preemptivly fix ColorPicker's test too
