+++ src/components/ColorPicker/tests/ColorPicker.test.tsx
@@ -156,40 +156,6 @@ describe('<ColorPicker />', () => {
 
       expect(event.preventDefault).not.toHaveBeenCalled();
     });
-
-    describe('isServer', () => {
-      jest.resetModules();
-      jest.mock('../../../utilities/target', () => ({
-        ...require.requireActual('../../../utilities/target'),
-        isServer: () => {
-          return true;
-        },
-      }));
-      const ColorPicker = require('../ColorPicker').default;
-      const Slidable = require('../components/Slidable').default;
-
-      it('does not attach the touchmove handler to the window when server side rendered', () => {
-        const colorPicker = mountWithAppProvider(
-          <ColorPicker color={red} onChange={noop} />,
-        );
-
-        colorPicker
-          .find(Slidable)
-          .first()
-          .simulate('mousedown');
-
-        const touch = {clientX: 0, clientY: 0};
-        const event = new TouchEvent('touchmove', {
-          touches: [touch],
-          cancelable: true,
-        } as TouchEventInit);
-        Object.assign(event, {preventDefault: jest.fn()});
-
-        window.dispatchEvent(event);
-
-        expect(event.preventDefault).toHaveBeenCalledTimes(1);
-      });
-    });
   });
 });
