+++ src/utilities/with-app-provider.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useContext} from 'react';
 import hoistStatics from 'hoist-non-react-statics';
 import {ClientApplication} from '@shopify/app-bridge';
 import {PolarisContext} from '../components/types';
@@ -52,66 +52,32 @@ export function withAppProvider<OwnProps>({withinScrollable}: Options = {}) {
   return function addProvider<C>(
     WrappedComponent: ReactComponent<OwnProps & WithAppProviderProps, C>,
   ): React.ComponentClass<OwnProps> & C {
-    // eslint-disable-next-line react/prefer-stateless-function
-    class WithProvider extends React.Component<OwnProps, never> {
-      static contextTypes = WrappedComponent.contextTypes;
+    const WithProvider: React.FunctionComponent = (props: OwnProps) => {
+      const link = useContext(LinkContext);
+      const theme = useContext(ThemeProviderContext);
+      const intl = useContext(I18nContext);
+      const scrollLockManager = useContext(ScrollLockManagerContext);
+      const stickyManager = useContext(StickyManagerContext);
+      const appBridge = useContext(AppBridgeContext);
 
-      render() {
-        return (
-          <LinkContext.Consumer>
-            {(link) => (
-              <ThemeProviderContext.Consumer>
-                {(theme) => (
-                  <I18nContext.Consumer>
-                    {(intl) => (
-                      <ScrollLockManagerContext.Consumer>
-                        {(scrollLockManager) => (
-                          <StickyManagerContext.Consumer>
-                            {(stickyManager) => (
-                              <AppBridgeContext.Consumer>
-                                {(appBridge) => {
-                                  const polarisContext: PolarisContext = {
-                                    link,
-                                    intl,
-                                    scrollLockManager,
-                                    stickyManager,
-                                    theme,
-                                    appBridge,
-                                  };
+      const polarisContext: PolarisContext = {
+        link,
+        intl,
+        scrollLockManager,
+        stickyManager,
+        theme,
+        appBridge,
+      };
 
-                                  if (
-                                    !intl ||
-                                    !scrollLockManager ||
-                                    !stickyManager ||
-                                    !link ||
-                                    !theme
-                                  ) {
-                                    throw new Error(
-                                      `The <AppProvider> component is required as of v2.0 of Polaris React. See https://polaris.shopify.com/components/structure/app-provider for implementation instructions.`,
-                                    );
-                                  }
-
-                                  return (
-                                    <WrappedComponent
-                                      {...this.props as any}
-                                      polaris={polarisContext}
-                                    />
-                                  );
-                                }}
-                              </AppBridgeContext.Consumer>
-                            )}
-                          </StickyManagerContext.Consumer>
-                        )}
-                      </ScrollLockManagerContext.Consumer>
-                    )}
-                  </I18nContext.Consumer>
-                )}
-              </ThemeProviderContext.Consumer>
-            )}
-          </LinkContext.Consumer>
+      if (!intl || !scrollLockManager || !stickyManager || !link || !theme) {
+        throw new Error(
+          `The <AppProvider> component is required as of v2.0 of Polaris React. See https://polaris.shopify.com/components/structure/app-provider for implementation instructions.`,
         );
       }
-    }
+
+      return <WrappedComponent {...props as any} polaris={polarisContext} />;
+    };
+    WithProvider.contextTypes = WrappedComponent.contextTypes;
 
     let WithScrollable: React.ComponentClass<any> | undefined;
     if (withinScrollable) {
